Crate freetype_sys

Modules

tt_tables 🔒

Structs

FT_BBox

FT_Bitmap

FT_BitmapGlyphRec

FT_Bitmap_Size

FT_CharMapRec

FT_Data

FT_FaceRec

FT_Generic

FT_GlyphRec

FT_GlyphSlotRec

FT_Glyph_Metrics

FT_LayerIterator

FT_ListNodeRec

FT_ListRec

FT_MM_Axis

FT_MM_Var

FT_Matrix

FT_MemoryRec

FT_Multi_Master

FT_Open_Args

FT_Outline

FT_OutlineGlyphRec

FT_Outline_Funcs

FT_Parameter

FT_Raster_Params

FT_SfntLangTag

FT_SfntName

FT_SizeRec

FT_Size_Metrics

FT_Size_RequestRec

FT_Span

FT_StreamRec

FT_UnitVector

FT_Var_Axis

FT_Var_Named_Style

FT_Vector

TT_OS2

TT_Postscript

Constants

FT_ENCODING_ADOBE_CUSTOM

FT_ENCODING_ADOBE_EXPERT

FT_ENCODING_ADOBE_LATIN_1

FT_ENCODING_ADOBE_STANDARD

FT_ENCODING_APPLE_ROMAN

FT_ENCODING_BIG5

FT_ENCODING_GB2312

FT_ENCODING_JOHAB

FT_ENCODING_MS_BIG5

FT_ENCODING_MS_GB2312

FT_ENCODING_MS_JOHAB

FT_ENCODING_MS_SJIS

FT_ENCODING_MS_SYMBOL

FT_ENCODING_MS_WANSUNG

FT_ENCODING_NONE

FT_ENCODING_OLD_LATIN_2

FT_ENCODING_SJIS

FT_ENCODING_UNICODE

FT_ENCODING_WANSUNG

FT_Err_Array_Too_Large

FT_Err_Bad_Argument

FT_Err_Bbx_Too_Big

FT_Err_CMap_Table_Missing

FT_Err_Cannot_Open_Resource

FT_Err_Cannot_Open_Stream

FT_Err_Cannot_Render_Glyph

FT_Err_Code_Overflow

FT_Err_Corrupted_Font_Glyphs

FT_Err_Corrupted_Font_Header

FT_Err_Could_Not_Find_Context

FT_Err_Debug_OpCode

FT_Err_Divide_By_Zero

FT_Err_ENDF_In_Exec_Stream

FT_Err_Execution_Too_Long

FT_Err_Hmtx_Table_Missing

FT_Err_Horiz_Header_Missing

FT_Err_Ignore

FT_Err_Invalid_Argument

FT_Err_Invalid_Cache_Handle

FT_Err_Invalid_CharMap_Format

FT_Err_Invalid_CharMap_Handle

FT_Err_Invalid_Character_Code

FT_Err_Invalid_CodeRange

FT_Err_Invalid_Composite

FT_Err_Invalid_Driver_Handle

FT_Err_Invalid_Face_Handle

FT_Err_Invalid_File_Format

FT_Err_Invalid_Frame_Operation

FT_Err_Invalid_Frame_Read

FT_Err_Invalid_Glyph_Format

FT_Err_Invalid_Glyph_Index

FT_Err_Invalid_Handle

FT_Err_Invalid_Horiz_Metrics

FT_Err_Invalid_Library_Handle

FT_Err_Invalid_Offset

FT_Err_Invalid_Opcode

FT_Err_Invalid_Outline

FT_Err_Invalid_PPem

FT_Err_Invalid_Pixel_Size

FT_Err_Invalid_Post_Table

FT_Err_Invalid_Post_Table_Format

FT_Err_Invalid_Reference

FT_Err_Invalid_Size_Handle

FT_Err_Invalid_Slot_Handle

FT_Err_Invalid_Stream_Handle

FT_Err_Invalid_Stream_Operation

FT_Err_Invalid_Stream_Read

FT_Err_Invalid_Stream_Seek

FT_Err_Invalid_Stream_Skip

FT_Err_Invalid_Table

FT_Err_Invalid_Version

FT_Err_Invalid_Vert_Metrics

FT_Err_Locations_Missing

FT_Err_Lower_Module_Version

FT_Err_Max

FT_Err_Missing_Bbx_Field

FT_Err_Missing_Chars_Field

FT_Err_Missing_Encoding_Field

FT_Err_Missing_Font_Field

FT_Err_Missing_Fontboundingbox_Field

FT_Err_Missing_Module

FT_Err_Missing_Property

FT_Err_Missing_Size_Field

FT_Err_Missing_Startchar_Field

FT_Err_Missing_Startfont_Field

FT_Err_Name_Table_Missing

FT_Err_Nested_DEFS

FT_Err_Nested_Frame_Access

FT_Err_No_Unicode_Glyph_Name

FT_Err_Ok

FT_Err_Out_Of_Memory

FT_Err_Post_Table_Missing

FT_Err_Raster_Corrupted

FT_Err_Raster_Negative_Height

FT_Err_Raster_Overflow

FT_Err_Raster_Uninitialized

FT_Err_Stack_Overflow

FT_Err_Stack_Underflow

FT_Err_Syntax_Error

FT_Err_Table_Missing

FT_Err_Too_Few_Arguments

FT_Err_Too_Many_Caches

FT_Err_Too_Many_Drivers

FT_Err_Too_Many_Extensions

FT_Err_Too_Many_Function_Defs

FT_Err_Too_Many_Hints

FT_Err_Too_Many_Instruction_Defs

FT_Err_Unimplemented_Feature

FT_Err_Unknown_File_Format

FT_Err_Unlisted_Object

FT_FACE_FLAG_CID_KEYED

FT_FACE_FLAG_COLOR

FT_FACE_FLAG_EXTERNAL_STREAM

FT_FACE_FLAG_FAST_GLYPHS

FT_FACE_FLAG_FIXED_SIZES

FT_FACE_FLAG_FIXED_WIDTH

FT_FACE_FLAG_GLYPH_NAMES

FT_FACE_FLAG_HINTER

FT_FACE_FLAG_HORIZONTAL

FT_FACE_FLAG_KERNING

FT_FACE_FLAG_MULTIPLE_MASTERS

FT_FACE_FLAG_SCALABLE

FT_FACE_FLAG_SFNT

FT_FACE_FLAG_TRICKY

FT_FACE_FLAG_VERTICAL

FT_FSTYPE_BITMAP_EMBEDDING_ONLY

FT_FSTYPE_EDITABLE_EMBEDDING

FT_FSTYPE_INSTALLABLE_EMBEDDING

FT_FSTYPE_NO_SUBSETTING

FT_FSTYPE_PREVIEW_AND_PRINT_EMBEDDING

FT_FSTYPE_RESTRICTED_LICENSE_EMBEDDING

FT_GLYPH_BBOX_GRIDFIT

FT_GLYPH_BBOX_PIXELS

FT_GLYPH_BBOX_SUBPIXELS

FT_GLYPH_BBOX_TRUNCATE

FT_GLYPH_BBOX_UNSCALED

FT_GLYPH_FORMAT_BITMAP

FT_GLYPH_FORMAT_COMPOSITE

FT_GLYPH_FORMAT_NONE

FT_GLYPH_FORMAT_OUTLINE

FT_GLYPH_FORMAT_PLOTTER

FT_KERNING_DEFAULT

FT_KERNING_UNFITTED

FT_KERNING_UNSCALED

FT_LCD_FILTER_DEFAULT

FT_LCD_FILTER_LEGACY

FT_LCD_FILTER_LEGACY1

FT_LCD_FILTER_LIGHT

FT_LCD_FILTER_MAX

FT_LCD_FILTER_NONE

FT_LOAD_COLOR

FT_LOAD_CROP_BITMAP

FT_LOAD_DEFAULT

FT_LOAD_FORCE_AUTOHINT

FT_LOAD_IGNORE_GLOBAL_ADVANCE_WIDTH

FT_LOAD_IGNORE_TRANSFORM

FT_LOAD_LINEAR_DESIGN

FT_LOAD_MONOCHROME

FT_LOAD_NO_AUTOHINT

FT_LOAD_NO_BITMAP

FT_LOAD_NO_HINTING

FT_LOAD_NO_RECURSE

FT_LOAD_NO_SCALE

FT_LOAD_PEDANTIC

FT_LOAD_RENDER

FT_LOAD_TARGET_LCD

FT_LOAD_TARGET_LCD_V

FT_LOAD_TARGET_LIGHT

FT_LOAD_TARGET_MONO

FT_LOAD_TARGET_NORMAL

FT_LOAD_VERTICAL_LAYOUT

FT_OPEN_DRIVER

FT_OPEN_MEMORY

FT_OPEN_PARAMS

FT_OPEN_PATHNAME

FT_OPEN_STREAM

FT_ORIENTATION_FILL_LEFT

FT_ORIENTATION_FILL_RIGHT

FT_ORIENTATION_NONE

FT_ORIENTATION_POSTSCRIPT

FT_ORIENTATION_TRUETYPE

FT_OUTLINE_EVEN_ODD_FILL

FT_OUTLINE_HIGH_PRECISION

FT_OUTLINE_IGNORE_DROPOUTS

FT_OUTLINE_INCLUDE_STUBS

FT_OUTLINE_NONE

FT_OUTLINE_OWNER

FT_OUTLINE_REVERSE_FILL

FT_OUTLINE_SINGLE_PASS

FT_OUTLINE_SMART_DROPOUTS

FT_PIXEL_MODE_BGRA

FT_PIXEL_MODE_GRAY

FT_PIXEL_MODE_GRAY2

FT_PIXEL_MODE_GRAY4

FT_PIXEL_MODE_LCD

FT_PIXEL_MODE_LCD_V

FT_PIXEL_MODE_MAX

FT_PIXEL_MODE_MONO

FT_PIXEL_MODE_NONE

FT_RENDER_MODE_LCD

FT_RENDER_MODE_LCD_V

FT_RENDER_MODE_LIGHT

FT_RENDER_MODE_MAX

FT_RENDER_MODE_MONO

FT_RENDER_MODE_NORMAL

FT_RENDER_MODE_SDF

FT_SFNT_HEAD

FT_SFNT_HHEA

FT_SFNT_MAX

FT_SFNT_MAXP

FT_SFNT_OS2

FT_SFNT_PCLT

FT_SFNT_POST

FT_SFNT_VHEA

FT_SIZE_REQUEST_TYPE_BBOX

FT_SIZE_REQUEST_TYPE_CELL

FT_SIZE_REQUEST_TYPE_MAX

FT_SIZE_REQUEST_TYPE_NOMINAL

FT_SIZE_REQUEST_TYPE_REAL_DIM

FT_SIZE_REQUEST_TYPE_SCALES

FT_STROKER_BORDER_LEFT

FT_STROKER_BORDER_RIGHT

FT_STROKER_LINECAP_BUTT

FT_STROKER_LINECAP_ROUND

FT_STROKER_LINECAP_SQUARE

FT_STROKER_LINEJOIN_BEVEL

FT_STROKER_LINEJOIN_MITER

FT_STROKER_LINEJOIN_MITER_FIXED

FT_STROKER_LINEJOIN_MITER_VARIABLE

FT_STROKER_LINEJOIN_ROUND

FT_STYLE_FLAG_BOLD

FT_STYLE_FLAG_ITALIC

FT_SUBGLYPH_FLAG_2X2

FT_SUBGLYPH_FLAG_ARGS_ARE_WORDS

FT_SUBGLYPH_FLAG_ARGS_ARE_XY_VALUES

FT_SUBGLYPH_FLAG_ROUND_XY_TO_GRID

FT_SUBGLYPH_FLAG_SCALE

FT_SUBGLYPH_FLAG_USE_MY_METRICS

FT_SUBGLYPH_FLAG_XY_SCALE

FT_VAR_AXIS_FLAG_HIDDEN

TT_ADOBE_ID_CUSTOM

TT_ADOBE_ID_EXPERT

TT_ADOBE_ID_LATIN_1

TT_ADOBE_ID_STANDARD

TT_APPLE_ID_DEFAULT

TT_APPLE_ID_FULL_UNICODE

TT_APPLE_ID_ISO_10646

TT_APPLE_ID_UNICODE_1_1

TT_APPLE_ID_UNICODE_2_0

TT_APPLE_ID_UNICODE_32

TT_APPLE_ID_VARIANT_SELECTOR

TT_ISO_ID_7BIT_ASCII

TT_ISO_ID_8859_1

TT_ISO_ID_10646

TT_MAC_ID_ARABIC

TT_MAC_ID_ARMENIAN

TT_MAC_ID_BENGALI

TT_MAC_ID_BURMESE

TT_MAC_ID_DEVANAGARI

TT_MAC_ID_GEEZ

TT_MAC_ID_GEORGIAN

TT_MAC_ID_GREEK

TT_MAC_ID_GUJARATI

TT_MAC_ID_GURMUKHI

TT_MAC_ID_HEBREW

TT_MAC_ID_JAPANESE

TT_MAC_ID_KANNADA

TT_MAC_ID_KHMER

TT_MAC_ID_KOREAN

TT_MAC_ID_LAOTIAN

TT_MAC_ID_MALAYALAM

TT_MAC_ID_MALDIVIAN

TT_MAC_ID_MONGOLIAN

TT_MAC_ID_ORIYA

TT_MAC_ID_ROMAN

TT_MAC_ID_RSYMBOL

TT_MAC_ID_RUSSIAN

TT_MAC_ID_SIMPLIFIED_CHINESE

TT_MAC_ID_SINDHI

TT_MAC_ID_SINHALESE

TT_MAC_ID_SLAVIC

TT_MAC_ID_TAMIL

TT_MAC_ID_TELUGU

TT_MAC_ID_THAI

TT_MAC_ID_TIBETAN

TT_MAC_ID_TRADITIONAL_CHINESE

TT_MAC_ID_UNINTERP

TT_MAC_ID_VIETNAMESE

TT_MAC_LANGID_AFRIKAANS

TT_MAC_LANGID_ALBANIAN

TT_MAC_LANGID_AMHARIC

TT_MAC_LANGID_ARABIC

TT_MAC_LANGID_ARMENIAN

TT_MAC_LANGID_ASSAMESE

TT_MAC_LANGID_AYMARA

TT_MAC_LANGID_AZERBAIJANI

TT_MAC_LANGID_AZERBAIJANI_ARABIC_SCRIPT

TT_MAC_LANGID_AZERBAIJANI_CYRILLIC_SCRIPT

TT_MAC_LANGID_AZERBAIJANI_ROMAN_SCRIPT

TT_MAC_LANGID_BASQUE

TT_MAC_LANGID_BENGALI

TT_MAC_LANGID_BRETON

TT_MAC_LANGID_BULGARIAN

TT_MAC_LANGID_BURMESE

TT_MAC_LANGID_BYELORUSSIAN

TT_MAC_LANGID_CATALAN

TT_MAC_LANGID_CHEWA

TT_MAC_LANGID_CHINESE_SIMPLIFIED

TT_MAC_LANGID_CHINESE_TRADITIONAL

TT_MAC_LANGID_CROATIAN

TT_MAC_LANGID_CZECH

TT_MAC_LANGID_DANISH

TT_MAC_LANGID_DUTCH

TT_MAC_LANGID_DZONGKHA

TT_MAC_LANGID_ENGLISH

TT_MAC_LANGID_ESPERANTO

TT_MAC_LANGID_ESTONIAN

TT_MAC_LANGID_FAEROESE

TT_MAC_LANGID_FARSI

TT_MAC_LANGID_FINNISH

TT_MAC_LANGID_FLEMISH

TT_MAC_LANGID_FRENCH

TT_MAC_LANGID_GALICIAN

TT_MAC_LANGID_GALLA

TT_MAC_LANGID_GEORGIAN

TT_MAC_LANGID_GERMAN

TT_MAC_LANGID_GREEK

TT_MAC_LANGID_GREEK_POLYTONIC

TT_MAC_LANGID_GREELANDIC

TT_MAC_LANGID_GUARANI

TT_MAC_LANGID_GUJARATI

TT_MAC_LANGID_HEBREW

TT_MAC_LANGID_HINDI

TT_MAC_LANGID_HUNGARIAN

TT_MAC_LANGID_ICELANDIC

TT_MAC_LANGID_INDONESIAN

TT_MAC_LANGID_INUKTITUT

TT_MAC_LANGID_IRISH

TT_MAC_LANGID_IRISH_GAELIC

TT_MAC_LANGID_ITALIAN

TT_MAC_LANGID_JAPANESE

TT_MAC_LANGID_JAVANESE

TT_MAC_LANGID_KANNADA

TT_MAC_LANGID_KASHMIRI

TT_MAC_LANGID_KAZAKH

TT_MAC_LANGID_KHMER

TT_MAC_LANGID_KIRGHIZ

TT_MAC_LANGID_KOREAN

TT_MAC_LANGID_KURDISH

TT_MAC_LANGID_LAO

TT_MAC_LANGID_LATIN

TT_MAC_LANGID_LETTISH

TT_MAC_LANGID_LITHUANIAN

TT_MAC_LANGID_MACEDONIAN

TT_MAC_LANGID_MALAGASY

TT_MAC_LANGID_MALAYALAM

TT_MAC_LANGID_MALAY_ARABIC_SCRIPT

TT_MAC_LANGID_MALAY_ROMAN_SCRIPT

TT_MAC_LANGID_MALTESE

TT_MAC_LANGID_MANX_GAELIC

TT_MAC_LANGID_MARATHI

TT_MAC_LANGID_MOLDAVIAN

TT_MAC_LANGID_MONGOLIAN

TT_MAC_LANGID_MONGOLIAN_CYRILLIC_SCRIPT

TT_MAC_LANGID_MONGOLIAN_MONGOLIAN_SCRIPT

TT_MAC_LANGID_NEPALI

TT_MAC_LANGID_NORWEGIAN

TT_MAC_LANGID_ORIYA

TT_MAC_LANGID_PASHTO

TT_MAC_LANGID_POLISH

TT_MAC_LANGID_PORTUGUESE

TT_MAC_LANGID_PUNJABI

TT_MAC_LANGID_QUECHUA

TT_MAC_LANGID_ROMANIAN

TT_MAC_LANGID_RUANDA

TT_MAC_LANGID_RUNDI

TT_MAC_LANGID_RUSSIAN

TT_MAC_LANGID_SAAMISK

TT_MAC_LANGID_SANSKRIT

TT_MAC_LANGID_SCOTTISH_GAELIC

TT_MAC_LANGID_SERBIAN

TT_MAC_LANGID_SINDHI

TT_MAC_LANGID_SINHALESE

TT_MAC_LANGID_SLOVAK

TT_MAC_LANGID_SLOVENIAN

TT_MAC_LANGID_SOMALI

TT_MAC_LANGID_SPANISH

TT_MAC_LANGID_SUNDANESE

TT_MAC_LANGID_SWAHILI

TT_MAC_LANGID_SWEDISH

TT_MAC_LANGID_TAGALOG

TT_MAC_LANGID_TAJIKI

TT_MAC_LANGID_TAMIL

TT_MAC_LANGID_TATAR

TT_MAC_LANGID_TELUGU

TT_MAC_LANGID_THAI

TT_MAC_LANGID_TIBETAN

TT_MAC_LANGID_TIGRINYA

TT_MAC_LANGID_TONGAN

TT_MAC_LANGID_TURKISH

TT_MAC_LANGID_TURKMEN

TT_MAC_LANGID_UIGHUR

TT_MAC_LANGID_UKRAINIAN

TT_MAC_LANGID_URDU

TT_MAC_LANGID_UZBEK

TT_MAC_LANGID_VIETNAMESE

TT_MAC_LANGID_WELSH

TT_MAC_LANGID_YIDDISH

TT_MS_ID_BIG_5

TT_MS_ID_JOHAB

TT_MS_ID_PRC

TT_MS_ID_SJIS

TT_MS_ID_SYMBOL_CS

TT_MS_ID_UCS_4

TT_MS_ID_UNICODE_CS

TT_MS_ID_WANSUNG

TT_MS_LANGID_AFRIKAANS_SOUTH_AFRICA

TT_MS_LANGID_ALBANIAN_ALBANIA

TT_MS_LANGID_ALSATIAN_FRANCE

TT_MS_LANGID_AMHARIC_ETHIOPIA

TT_MS_LANGID_ARABIC_ALGERIA

TT_MS_LANGID_ARABIC_BAHRAIN

TT_MS_LANGID_ARABIC_EGYPT

TT_MS_LANGID_ARABIC_IRAQ

TT_MS_LANGID_ARABIC_JORDAN

TT_MS_LANGID_ARABIC_KUWAIT

TT_MS_LANGID_ARABIC_LEBANON

TT_MS_LANGID_ARABIC_LIBYA

TT_MS_LANGID_ARABIC_MOROCCO

TT_MS_LANGID_ARABIC_OMAN

TT_MS_LANGID_ARABIC_QATAR

TT_MS_LANGID_ARABIC_SAUDI_ARABIA

TT_MS_LANGID_ARABIC_SYRIA

TT_MS_LANGID_ARABIC_TUNISIA

TT_MS_LANGID_ARABIC_UAE

TT_MS_LANGID_ARABIC_YEMEN

TT_MS_LANGID_ARMENIAN_ARMENIA

TT_MS_LANGID_ASSAMESE_INDIA

TT_MS_LANGID_AZERI_AZERBAIJAN_CYRILLIC

TT_MS_LANGID_AZERI_AZERBAIJAN_LATIN

TT_MS_LANGID_BASHKIR_RUSSIA

TT_MS_LANGID_BASQUE_BASQUE

TT_MS_LANGID_BELARUSIAN_BELARUS

TT_MS_LANGID_BENGALI_BANGLADESH

TT_MS_LANGID_BENGALI_INDIA

TT_MS_LANGID_BOSNIAN_BOSNIA_HERZEGOVINA

TT_MS_LANGID_BOSNIAN_BOSNIA_HERZ_CYRILLIC

TT_MS_LANGID_BRETON_FRANCE

TT_MS_LANGID_BULGARIAN_BULGARIA

TT_MS_LANGID_CATALAN_CATALAN

TT_MS_LANGID_CHINESE_HONG_KONG

TT_MS_LANGID_CHINESE_MACAO

TT_MS_LANGID_CHINESE_PRC

TT_MS_LANGID_CHINESE_SINGAPORE

TT_MS_LANGID_CHINESE_TAIWAN

TT_MS_LANGID_CORSICAN_FRANCE

TT_MS_LANGID_CROATIAN_BOSNIA_HERZEGOVINA

TT_MS_LANGID_CROATIAN_CROATIA

TT_MS_LANGID_CZECH_CZECH_REPUBLIC

TT_MS_LANGID_DANISH_DENMARK

TT_MS_LANGID_DARI_AFGHANISTAN

TT_MS_LANGID_DHIVEHI_MALDIVES

TT_MS_LANGID_DUTCH_BELGIUM

TT_MS_LANGID_DUTCH_NETHERLANDS

TT_MS_LANGID_ENGLISH_AUSTRALIA

TT_MS_LANGID_ENGLISH_BELIZE

TT_MS_LANGID_ENGLISH_CANADA

TT_MS_LANGID_ENGLISH_CARIBBEAN

TT_MS_LANGID_ENGLISH_INDIA

TT_MS_LANGID_ENGLISH_IRELAND

TT_MS_LANGID_ENGLISH_JAMAICA

TT_MS_LANGID_ENGLISH_MALAYSIA

TT_MS_LANGID_ENGLISH_NEW_ZEALAND

TT_MS_LANGID_ENGLISH_PHILIPPINES

TT_MS_LANGID_ENGLISH_SINGAPORE

TT_MS_LANGID_ENGLISH_SOUTH_AFRICA

TT_MS_LANGID_ENGLISH_TRINIDAD

TT_MS_LANGID_ENGLISH_UNITED_KINGDOM

TT_MS_LANGID_ENGLISH_UNITED_STATES

TT_MS_LANGID_ENGLISH_ZIMBABWE

TT_MS_LANGID_ESTONIAN_ESTONIA

TT_MS_LANGID_FAEROESE_FAEROE_ISLANDS

TT_MS_LANGID_FILIPINO_PHILIPPINES

TT_MS_LANGID_FINNISH_FINLAND

TT_MS_LANGID_FRENCH_BELGIUM

TT_MS_LANGID_FRENCH_CANADA

TT_MS_LANGID_FRENCH_FRANCE

TT_MS_LANGID_FRENCH_LUXEMBOURG

TT_MS_LANGID_FRENCH_MONACO

TT_MS_LANGID_FRENCH_SWITZERLAND

TT_MS_LANGID_FRISIAN_NETHERLANDS

TT_MS_LANGID_GALICIAN_GALICIAN

TT_MS_LANGID_GEORGIAN_GEORGIA

TT_MS_LANGID_GERMAN_AUSTRIA

TT_MS_LANGID_GERMAN_GERMANY

TT_MS_LANGID_GERMAN_LIECHTENSTEIN

TT_MS_LANGID_GERMAN_LUXEMBOURG

TT_MS_LANGID_GERMAN_SWITZERLAND

TT_MS_LANGID_GREEK_GREECE

TT_MS_LANGID_GREENLANDIC_GREENLAND

TT_MS_LANGID_GUJARATI_INDIA

TT_MS_LANGID_HAUSA_NIGERIA

TT_MS_LANGID_HEBREW_ISRAEL

TT_MS_LANGID_HINDI_INDIA

TT_MS_LANGID_HUNGARIAN_HUNGARY

TT_MS_LANGID_ICELANDIC_ICELAND

TT_MS_LANGID_IGBO_NIGERIA

TT_MS_LANGID_INDONESIAN_INDONESIA

TT_MS_LANGID_INUKTITUT_CANADA

TT_MS_LANGID_INUKTITUT_CANADA_LATIN

TT_MS_LANGID_IRISH_IRELAND

TT_MS_LANGID_ISIXHOSA_SOUTH_AFRICA

TT_MS_LANGID_ISIZULU_SOUTH_AFRICA

TT_MS_LANGID_ITALIAN_ITALY

TT_MS_LANGID_ITALIAN_SWITZERLAND

TT_MS_LANGID_JAPANESE_JAPAN

TT_MS_LANGID_KANNADA_INDIA

TT_MS_LANGID_KAZAKH_KAZAKHSTAN

TT_MS_LANGID_KHMER_CAMBODIA

TT_MS_LANGID_KICHE_GUATEMALA

TT_MS_LANGID_KINYARWANDA_RWANDA

TT_MS_LANGID_KISWAHILI_KENYA

TT_MS_LANGID_KONKANI_INDIA

TT_MS_LANGID_KOREAN_KOREA

TT_MS_LANGID_KYRGYZ_KYRGYZSTAN

TT_MS_LANGID_LAO_LAOS

TT_MS_LANGID_LATVIAN_LATVIA

TT_MS_LANGID_LITHUANIAN_LITHUANIA

TT_MS_LANGID_LOWER_SORBIAN_GERMANY

TT_MS_LANGID_LUXEMBOURGISH_LUXEMBOURG

TT_MS_LANGID_MACEDONIAN_MACEDONIA

TT_MS_LANGID_MALAYALAM_INDIA

TT_MS_LANGID_MALAY_BRUNEI_DARUSSALAM

TT_MS_LANGID_MALAY_MALAYSIA

TT_MS_LANGID_MALTESE_MALTA

TT_MS_LANGID_MAORI_NEW_ZEALAND

TT_MS_LANGID_MAPUDUNGUN_CHILE

TT_MS_LANGID_MARATHI_INDIA

TT_MS_LANGID_MOHAWK_MOHAWK

TT_MS_LANGID_MONGOLIAN_MONGOLIA

TT_MS_LANGID_MONGOLIAN_PRC

TT_MS_LANGID_NEPALI_NEPAL

TT_MS_LANGID_NORWEGIAN_NORWAY_BOKMAL

TT_MS_LANGID_NORWEGIAN_NORWAY_NYNORSK

TT_MS_LANGID_OCCITAN_FRANCE

TT_MS_LANGID_ODIA_INDIA

TT_MS_LANGID_PASHTO_AFGHANISTAN

TT_MS_LANGID_POLISH_POLAND

TT_MS_LANGID_PORTUGUESE_BRAZIL

TT_MS_LANGID_PORTUGUESE_PORTUGAL

TT_MS_LANGID_PUNJABI_INDIA

TT_MS_LANGID_QUECHUA_BOLIVIA

TT_MS_LANGID_QUECHUA_ECUADOR

TT_MS_LANGID_QUECHUA_PERU

TT_MS_LANGID_ROMANIAN_ROMANIA

TT_MS_LANGID_ROMANSH_SWITZERLAND

TT_MS_LANGID_RUSSIAN_RUSSIA

TT_MS_LANGID_SAMI_INARI_FINLAND

TT_MS_LANGID_SAMI_LULE_NORWAY

TT_MS_LANGID_SAMI_LULE_SWEDEN

TT_MS_LANGID_SAMI_NORTHERN_FINLAND

TT_MS_LANGID_SAMI_NORTHERN_NORWAY

TT_MS_LANGID_SAMI_NORTHERN_SWEDEN

TT_MS_LANGID_SAMI_SKOLT_FINLAND

TT_MS_LANGID_SAMI_SOUTHERN_NORWAY

TT_MS_LANGID_SAMI_SOUTHERN_SWEDEN

TT_MS_LANGID_SANSKRIT_INDIA

TT_MS_LANGID_SERBIAN_BOSNIA_HERZ_CYRILLIC

TT_MS_LANGID_SERBIAN_BOSNIA_HERZ_LATIN

TT_MS_LANGID_SERBIAN_SERBIA_CYRILLIC

TT_MS_LANGID_SERBIAN_SERBIA_LATIN

TT_MS_LANGID_SESOTHO_SA_LEBOA_SOUTH_AFRICA

TT_MS_LANGID_SETSWANA_SOUTH_AFRICA

TT_MS_LANGID_SINHALA_SRI_LANKA

TT_MS_LANGID_SLOVAK_SLOVAKIA

TT_MS_LANGID_SLOVENIAN_SLOVENIA

TT_MS_LANGID_SPANISH_ARGENTINA

TT_MS_LANGID_SPANISH_BOLIVIA

TT_MS_LANGID_SPANISH_CHILE

TT_MS_LANGID_SPANISH_COLOMBIA

TT_MS_LANGID_SPANISH_COSTA_RICA

TT_MS_LANGID_SPANISH_DOMINICAN_REPUBLIC

TT_MS_LANGID_SPANISH_ECUADOR

TT_MS_LANGID_SPANISH_EL_SALVADOR

TT_MS_LANGID_SPANISH_GUATEMALA

TT_MS_LANGID_SPANISH_HONDURAS

TT_MS_LANGID_SPANISH_MEXICO

TT_MS_LANGID_SPANISH_NICARAGUA

TT_MS_LANGID_SPANISH_PANAMA

TT_MS_LANGID_SPANISH_PARAGUAY

TT_MS_LANGID_SPANISH_PERU

TT_MS_LANGID_SPANISH_PUERTO_RICO

TT_MS_LANGID_SPANISH_SPAIN_MODERN_SORT

TT_MS_LANGID_SPANISH_SPAIN_TRADITIONAL_SORT

TT_MS_LANGID_SPANISH_UNITED_STATES

TT_MS_LANGID_SPANISH_URUGUAY

TT_MS_LANGID_SPANISH_VENEZUELA

TT_MS_LANGID_SWEDISH_FINLAND

TT_MS_LANGID_SWEDISH_SWEDEN

TT_MS_LANGID_SYRIAC_SYRIA

TT_MS_LANGID_TAJIK_TAJIKISTAN

TT_MS_LANGID_TAMAZIGHT_ALGERIA

TT_MS_LANGID_TAMIL_INDIA

TT_MS_LANGID_TATAR_RUSSIA

TT_MS_LANGID_TELUGU_INDIA

TT_MS_LANGID_THAI_THAILAND

TT_MS_LANGID_TIBETAN_PRC

TT_MS_LANGID_TURKISH_TURKEY

TT_MS_LANGID_TURKMEN_TURKMENISTAN

TT_MS_LANGID_UIGHUR_PRC

TT_MS_LANGID_UKRAINIAN_UKRAINE

TT_MS_LANGID_UPPER_SORBIAN_GERMANY

TT_MS_LANGID_URDU_PAKISTAN

TT_MS_LANGID_UZBEK_UZBEKISTAN_CYRILLIC

TT_MS_LANGID_UZBEK_UZBEKISTAN_LATIN

TT_MS_LANGID_VIETNAMESE_VIET_NAM

TT_MS_LANGID_WELSH_UNITED_KINGDOM

TT_MS_LANGID_WOLOF_SENEGAL

TT_MS_LANGID_YAKUT_RUSSIA

TT_MS_LANGID_YI_PRC

TT_MS_LANGID_YORUBA_NIGERIA

TT_NAME_ID_CID_FINDFONT_NAME

TT_NAME_ID_COPYRIGHT

TT_NAME_ID_DARK_BACKGROUND

TT_NAME_ID_DESCRIPTION

TT_NAME_ID_DESIGNER

TT_NAME_ID_DESIGNER_URL

TT_NAME_ID_FONT_FAMILY

TT_NAME_ID_FONT_SUBFAMILY

TT_NAME_ID_FULL_NAME

TT_NAME_ID_LICENSE

TT_NAME_ID_LICENSE_URL

TT_NAME_ID_LIGHT_BACKGROUND

TT_NAME_ID_MAC_FULL_NAME

TT_NAME_ID_MANUFACTURER

TT_NAME_ID_PS_NAME

TT_NAME_ID_SAMPLE_TEXT

TT_NAME_ID_TRADEMARK

TT_NAME_ID_TYPOGRAPHIC_FAMILY

TT_NAME_ID_TYPOGRAPHIC_SUBFAMILY

TT_NAME_ID_UNIQUE_ID

TT_NAME_ID_VARIATIONS_PREFIX

TT_NAME_ID_VENDOR_URL

TT_NAME_ID_VERSION_STRING

TT_NAME_ID_WWS_FAMILY

TT_NAME_ID_WWS_SUBFAMILY

TT_PLATFORM_ADOBE

TT_PLATFORM_APPLE_UNICODE

TT_PLATFORM_CUSTOM

TT_PLATFORM_ISO

TT_PLATFORM_MACINTOSH

TT_PLATFORM_MICROSOFT

ft_sfnt_head

ft_sfnt_hhea

ft_sfnt_max

ft_sfnt_maxp

ft_sfnt_os2

ft_sfnt_pclt

ft_sfnt_post

ft_sfnt_vhea

Traits

FTErrorMethods

Functions

FT_Add_Default_Modules⚠

FT_Attach_File⚠

FT_Attach_Stream⚠

FT_CeilFix⚠

FT_DivFix⚠

FT_Done_Face⚠

FT_Done_FreeType⚠

FT_Done_Glyph⚠

FT_Done_Library⚠

FT_Done_MM_Var⚠

FT_Face_GetCharVariantIndex⚠

FT_Face_GetCharVariantIsDefault⚠

FT_Face_GetCharsOfVariant⚠

FT_Face_GetVariantSelectors⚠

FT_Face_GetVariantsOfChar⚠

FT_FloorFix⚠

FT_Get_Char_Index⚠

FT_Get_Charmap_Index⚠

FT_Get_Color_Glyph_Layer⚠

FT_Get_FSType_Flags⚠

FT_Get_First_Char⚠

FT_Get_Glyph⚠

FT_Get_Glyph_Name⚠

FT_Get_Kerning⚠

FT_Get_MM_Blend_Coordinates⚠

FT_Get_MM_Var⚠

FT_Get_MM_WeightVector⚠

FT_Get_Multi_Master⚠

FT_Get_Name_Index⚠

FT_Get_Next_Char⚠

FT_Get_Postscript_Name⚠

FT_Get_Sfnt_LangTag⚠

FT_Get_Sfnt_Name⚠

FT_Get_Sfnt_Name_Count⚠

FT_Get_Sfnt_Table⚠

FT_Get_SubGlyph_Info⚠

FT_Get_Track_Kerning⚠

FT_Get_Var_Axis_Flags⚠

FT_Get_Var_Blend_Coordinates⚠

FT_Get_Var_Design_Coordinates⚠

FT_GlyphSlot_Embolden⚠

FT_GlyphSlot_Oblique⚠

FT_Glyph_Copy⚠

FT_Glyph_Get_CBox⚠

FT_Glyph_Stroke⚠

FT_Glyph_StrokeBorder⚠

FT_Glyph_To_Bitmap⚠

FT_Glyph_Transform⚠

FT_HAS_COLOR⚠

Returns true whenever a face object contains tables for color glyphs.

FT_HAS_FIXED_SIZES⚠

Returns true whenever a face object contains some embedded bitmaps.

FT_HAS_GLYPH_NAMES⚠

Returns true whenever a face object contains some glyph names that can be accessed through FT_Get_Glyph_Name().

FT_HAS_HORIZONTAL⚠

Returns true whenever a face object contains horizontal metrics.

FT_HAS_KERNING⚠

Returns true whenever a face object contains kerning data that can be accessed with FT_Get_Kerning().

FT_HAS_MULTIPLE_MASTERS⚠

Returns true whenever a face object contains some multiple masters.

FT_HAS_VERTICAL⚠

Returns true whenever a face object contains real vertical metrics (and not only synthesized ones).

FT_IS_CID_KEYED⚠

Returns true whenever a face object contains a CID-keyed font.

FT_IS_FIXED_WIDTH⚠

Returns true whenever a face object contains a font face that contains fixed-width (or ‘monospace’, ‘fixed-pitch’, etc.) glyphs.

FT_IS_SCALABLE⚠

Returns true whenever a face object contains a scalable font face.

FT_IS_SFNT⚠

Returns true whenever a face object contains a font whose format is based on the SFNT storage scheme.

FT_IS_TRICKY⚠

Returns true whenever a face represents a ‘tricky’ font.

FT_Init_FreeType⚠

FT_Library_SetLcdFilter⚠

FT_Load_Char⚠

FT_Load_Glyph⚠

FT_Load_Sfnt_Table⚠

FT_MulDiv⚠

FT_MulFix⚠

FT_New_Face⚠

FT_New_Library⚠

FT_New_Memory_Face⚠

FT_Open_Face⚠

FT_Outline_Check⚠

FT_Outline_Copy⚠

FT_Outline_Decompose⚠

FT_Outline_Done⚠

FT_Outline_Embolden⚠

FT_Outline_EmboldenXY⚠

FT_Outline_GetInsideBorder⚠

FT_Outline_GetOutsideBorder⚠

FT_Outline_Get_BBox⚠

FT_Outline_Get_Bitmap⚠

FT_Outline_Get_CBox⚠

FT_Outline_Get_Orientation⚠

FT_Outline_New⚠

FT_Outline_Render⚠

FT_Outline_Reverse⚠

FT_Outline_Transform⚠

FT_Outline_Translate⚠

FT_Property_Get⚠

FT_Property_Set⚠

FT_Reference_Face⚠

FT_Reference_Library⚠

FT_Render_Glyph⚠

FT_Request_Size⚠

FT_RoundFix⚠

FT_Select_Charmap⚠

FT_Select_Size⚠

FT_Set_Char_Size⚠

FT_Set_Charmap⚠

FT_Set_Default_Properties⚠

FT_Set_MM_Blend_Coordinates⚠

FT_Set_MM_Design_Coordinates⚠

FT_Set_MM_WeightVector⚠

FT_Set_Named_Instance⚠

FT_Set_Pixel_Sizes⚠

FT_Set_Transform⚠

FT_Set_Var_Blend_Coordinates⚠

FT_Set_Var_Design_Coordinates⚠

FT_Stroker_BeginSubPath⚠

FT_Stroker_ConicTo⚠

FT_Stroker_CubicTo⚠

FT_Stroker_Done⚠

FT_Stroker_EndSubPath⚠

FT_Stroker_Export⚠

FT_Stroker_ExportBorder⚠

FT_Stroker_GetBorderCounts⚠

FT_Stroker_GetCounts⚠

FT_Stroker_LineTo⚠

FT_Stroker_New⚠

FT_Stroker_ParseOutline⚠

FT_Stroker_Rewind⚠

FT_Stroker_Set⚠

Type Aliases

FT_Alloc_Func

FT_BitmapGlyph

FT_Bool

FT_Byte

FT_Bytes

FT_Char

FT_CharMap

FT_Driver

FT_DriverRec

FT_Encoding

FT_Error

FT_F2Dot14

FT_F26Dot6

FT_FWord

FT_Face

FT_Face_Internal

FT_Face_InternalRec

FT_Fixed

FT_Free_Func

FT_Generic_Finalizer

FT_Glyph

FT_GlyphSlot

FT_Glyph_BBox_Mode

FT_Glyph_Format

FT_Int

FT_Int16

FT_Int32

FT_Int64

FT_Kerning_Mode

FT_LcdFilter

FT_Library

FT_LibraryRec

FT_ListNode

FT_Long

FT_Memory

FT_Module

FT_ModuleRec

FT_Offset

FT_Orientation

FT_OutlineGlyph

FT_Outline_ConicToFunc

FT_Outline_CubicToFunc

FT_Outline_LineToFunc

FT_Outline_MoveToFunc

FT_Pixel_Mode

FT_Pointer

FT_Pos

FT_PtrDist

FT_Raster_BitSet_Func

FT_Raster_BitTest_Func

FT_Realloc_Func

FT_Render_Mode

FT_Renderer

FT_RendererRec

FT_Sfnt_Tag

FT_Short

FT_Size

FT_Size_Internal

FT_Size_InternalRec

FT_Size_Request

FT_Size_Request_Type

FT_Slot_Internal

FT_Slot_InternalRec

FT_SpanFunc

FT_Stream

FT_StreamDesc

FT_Stream_CloseFunc

FT_Stream_IoFunc

FT_String

FT_Stroker

FT_StrokerBorder

FT_StrokerRec

FT_Stroker_LineCap

FT_Stroker_LineJoin

FT_SubGlyph

FT_SubGlyphRec

FT_Tag

FT_UFWord

FT_UInt

FT_UInt16

FT_UInt32

FT_UInt64

FT_ULong

FT_UShort

TT_OS2_Internal

TT_Postscript_Internal

enum_FT_Sfnt_Tag_